+++ CHANGELOG.md
@@ -2,8 +2,8 @@
 
 ## [Unreleased]
 ### Added
-- Data ingress support to GlusterFS. Please see the configuration doc for
-more information.
+- Data ingress support to GlusterFS and Azure Blob Storage. Please see the
+configuration doc for more information.
 - Experimental support for OpenSSH HPN on Ubuntu
 - Additional actions: `ingressdata`, `gettaskallfiles`, `listjobs`,
 `listtasks`. Please see the usage doc for more information.